requirements.txt
python=3.8
pip>=23.0
fastapi
uvicorn
scrapy
selenium
beautifulsoup4
requests
aiohttp
sqlalchemy
psycopg2
pillow
pydantic
python-jose
passlib
python-dateutil
pytest
pytest-asyncio
pytest-cov
flake8
mypy
isort